Pin postgres version. Set postgres user/pass.
A minor version upgrade made a backwards-incompatible change which broke
CI/CD and competitor use of server image. Pins the postgres image to a
specific version to prevent this from happening in the future.

Sets the username and password for the database superuser account, which
is now required to connect. Note for security this is preferrable to
POSTGRES_HOST_AUTH_METHOD, which would accept any connection that can
access the port, as we can change the password at competition to lockout
teams, similar to how we change the session tokens.